+++ src/background/deploymentUpdater.test.ts
@@ -34,10 +34,7 @@ import {
   getSettingsState,
   saveSettingsState,
 } from "@/store/settings/settingsStorage";
-import {
-  getEditorState,
-  saveEditorState,
-} from "@/pageEditor/store/editor/editorStorage";
+import { getEditorState, saveEditorState } from "@/store/editorStorage";
 import {
   editorSlice,
   initialState as initialEditorState,
