+++ src/pageEditor/store/editor/editorSliceHelpers.test.ts
@@ -23,7 +23,15 @@ import {
   makeInitialBrickConfigurationUIState,
 } from "@/pageEditor/store/editor/uiState";
 import { getPipelineMap } from "@/pageEditor/tabs/editTab/editHelpers";
-
+import {
+  ensureBrickPipelineUIState,
+  ensureBrickConfigurationUIState,
+  removeModComponentFormState,
+  removeModData,
+  setActiveModId,
+  setActiveNodeId,
+  syncBrickConfigurationUIStates,
+} from "@/pageEditor/store/editor/editorSliceHelpers";
 import { produce } from "immer";
 import {
   type BrickPipelineUIState,
@@ -42,13 +50,6 @@ import {
 } from "@/pageEditor/store/editor/editorSelectors";
 import { modMetadataFactory } from "@/testUtils/factories/modComponentFactories";
 import { formStateFactory } from "@/testUtils/factories/pageEditorFactories";
-import { ensureBrickConfigurationUIState } from "@/pageEditor/store/editor/editorSliceHelpers/ensureBrickConfigurationUiState";
-import { ensureBrickPipelineUIState } from "@/pageEditor/store/editor/editorSliceHelpers/ensureBrickPipelineUiState";
-import { removeModComponentFormState } from "@/pageEditor/store/editor/editorSliceHelpers/removeModComponentFormState";
-import { removeModData } from "@/pageEditor/store/editor/editorSliceHelpers/removeModData";
-import { setActiveModId } from "@/pageEditor/store/editor/editorSliceHelpers/setActiveModId";
-import { setActiveNodeId } from "@/pageEditor/store/editor/editorSliceHelpers/setActiveNodeId";
-import { syncBrickConfigurationUIStates } from "@/pageEditor/store/editor/editorSliceHelpers/syncBrickConfigurationUiStates";
 
 describe("ensureBrickPipelineUIState", () => {
   test("does not affect existing ui state", () => {
