--- src/pageEditor/store/editor/editorSliceHelpers/editModMetadata.ts
@@ -1,35 +1,0 @@
-/*
- * Copyright (C) 2024 PixieBrix, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import {
-  type EditorState,
-  type ModMetadataFormState,
-} from "@/pageEditor/store/editor/pageEditorTypes";
-import { type Draft } from "immer";
-
-export function editModMetadata(
-  state: Draft<EditorState>,
-  metadata: ModMetadataFormState,
-) {
-  const { activeModId } = state;
-  if (activeModId == null) {
-    return;
-  }
-
-  // eslint-disable-next-line security/detect-object-injection -- not user input
-  state.dirtyModMetadataById[activeModId] = metadata;
-}
