--- src/pageEditor/store/editor/editorSliceHelpers/setActiveModComponentId.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2024 PixieBrix, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { type Draft } from "immer";
-import { type EditorState } from "@/pageEditor/store/editor/pageEditorTypes";
-import { type ModComponentFormState } from "@/pageEditor/starterBricks/formStateTypes";
-import { ensureBrickPipelineUIState } from "@/pageEditor/store/editor/editorSliceHelpers/ensureBrickPipelineUiState";
-
-export function setActiveModComponentId(
-  state: Draft<EditorState>,
-  modComponentFormState: ModComponentFormState,
-) {
-  state.error = null;
-  state.beta = false;
-  state.activeModComponentId = modComponentFormState.uuid;
-  state.activeModId = null;
-  state.expandedModId =
-    modComponentFormState.modMetadata?.id ?? state.expandedModId;
-  state.selectionSeq++;
-
-  ensureBrickPipelineUIState(state, modComponentFormState.uuid);
-}
