--- src/pageEditor/store/editor/editorSliceHelpers/setActiveModId.ts
@@ -1,36 +1,0 @@
-/*
- * Copyright (C) 2024 PixieBrix, Inc.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-import { type EditorState } from "@/pageEditor/store/editor/pageEditorTypes";
-import { type RegistryId } from "@/types/registryTypes";
-import { type Draft } from "immer";
-
-export function setActiveModId(state: Draft<EditorState>, modId: RegistryId) {
-  state.error = null;
-  state.beta = false;
-  state.activeModComponentId = null;
-
-  if (state.expandedModId === modId && state.activeModId === modId) {
-    // "un-toggle" the mod, if it's already selected
-    state.expandedModId = null;
-  } else {
-    state.expandedModId = modId;
-  }
-
-  state.activeModId = modId;
-  state.selectionSeq++;
-}
