+++ src/tsconfig.strictNullChecks.json
@@ -1158,16 +1158,7 @@
     "./pageEditor/modListingPanel/modals/SaveAsNewModModal.tsx",
     "./pageEditor/store/editor/editorSelectors.ts",
     "./pageEditor/store/editor/editorSlice.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/editModMetadata.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/editModOptionsDefinitions.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/ensureBrickConfigurationUiState.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/ensureBrickPipelineUiState.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/removeModComponentFormState.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/removeModData.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/setActiveModComponentId.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/setActiveModId.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/setActiveNodeId.ts",
-    "./pageEditor/store/editor/editorSliceHelpers/syncBrickConfigurationUiStates.ts",
+    "./pageEditor/store/editor/editorSliceHelpers.ts",
     "./pageEditor/store/runtime/runtimeSelectors.ts",
     "./pageEditor/store/runtime/runtimeSlice.ts",
     "./pageEditor/store/runtime/runtimeSliceTypes.ts",
@@ -1399,9 +1390,9 @@
     "./store/commonActions.ts",
     "./store/deactivateUtils.ts",
     "./store/defaultMiddlewareConfig.ts",
-    "./pageEditor/store/editor/editorMigrations.test.ts",
-    "./pageEditor/store/editor/editorMigrations.ts",
-    "./pageEditor/store/editor/editorStorage.ts",
+    "./store/editorMigrations.test.ts",
+    "./store/editorMigrations.ts",
+    "./store/editorStorage.ts",
     "./store/enterprise/managedStorage.test.ts",
     "./store/enterprise/managedStorage.ts",
     "./store/enterprise/managedStorageTypes.ts",
